fix: subscriptions to their own module
